Minimum viable local taxon, comment, id
