+++ tests/factories/LocalComment.js
@@ -1,9 +1,5 @@
-import factory, { define } from "factoria";
+import { define } from "factoria";
 
 export default define( "LocalComment", faker => ( {
-  uuid: faker.datatype.uuid( ),
-  id: faker.datatype.number( ),
-  createdAt: "2019-09-09T08:28:05-08:00",
-  user: factory( "RemoteUser" ),
-  body: faker.lorem.sentence( )
+  uuid: faker.datatype.uuid( )
 } ) );
