+++ tests/factories/LocalObservation.js
@@ -16,7 +16,13 @@ export default define( "LocalObservation", faker => ( {
     factory( "LocalObservationPhoto" )
   ],
   placeGuess: "SF",
-  taxon: factory( "LocalTaxon" ),
+  taxon: factory( "LocalTaxon", {
+    name: faker.name.firstName( ),
+    rank: "species",
+    rank_level: 10,
+    preferred_common_name: faker.name.fullName( ),
+    defaultPhoto: factory( "LocalPhoto" )
+  } ),
   user: factory( "LocalUser" ),
   qualityGrade: "research",
   latitude: Number( faker.address.latitude( ) ),
@@ -27,5 +33,7 @@ export default define( "LocalObservation", faker => ( {
   wasSynced: jest.fn( ),
   observed_on_string: "2022-12-03T11:14:16",
   // This is a Realm object method that we use to see if a record was deleted or not
-  isValid: jest.fn( () => true )
+  isValid: jest.fn( () => true ),
+  comments_viewed: false,
+  identifications_view: false
 } ) );
